# public/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/rooms": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get list of all rooms with associated user and game details, and available games",
                "operationId": "fc04f85434400dbf0e26d2ec77a5878d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rooms": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Room"
                                            }
                                        },
                                        "games": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Game"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Room"
                ],
                "summary": "Create a new room",
                "operationId": "afb43684c849533c65e931df636b4de6",
                "requestBody": {
                    "description": "Provide room details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "integer",
                                        "example": 1234
                                    },
                                    "type": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "optional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/rooms/{id}": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get a single room with associated user and game",
                "operationId": "1c71dfd30793f918f07dc95b23daefc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Room"
                ],
                "summary": "Update an existing room",
                "operationId": "ea75b11760a871bba18cf5f9e43a5b70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Room data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Room"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Room"
                ],
                "summary": "Delete a room",
                "operationId": "cf9104dbefd91a81bebb07cafe7e70cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Room deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Game": {
                "title": "Game",
                "description": "Game model",
                "properties": {
                    "id": {
                        "description": "ID of the game",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "description": "Name of the game",
                        "type": "string",
                        "example": "Chess"
                    }
                },
                "type": "object"
            },
            "Room": {
                "title": "Room",
                "description": "Room model",
                "properties": {
                    "id": {
                        "description": "ID of the room",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "description": "Name of the room",
                        "type": "integer",
                        "example": 1234
                    },
                    "type": {
                        "description": "Type of the room",
                        "type": "boolean",
                        "example": true
                    },
                    "password": {
                        "description": "Password of the room",
                        "type": "string",
                        "example": "optional"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "game": {
                        "$ref": "#/components/schemas/Game"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "description": "Username",
                        "type": "string",
                        "example": "johndoe"
                    },
                    "password": {
                        "description": "Password",
                        "type": "string",
                        "example": "pass1234"
                    }
                },
                "type": "object"
            }
        }
    }
}